refactor: 重构 MaaBuilder
