--- .maabundlerignore
@@ -1,8 +1,0 @@
-libiomp5md.dll
-mkldnn.dll
-mklml.dll
-opencv_world453.dll
-paddle_inference.dll
-ppocr.dll
-aria2c.exe
-resource/PaddleOCR/**
