+++ .nuke/build.schema.json
@@ -66,20 +66,15 @@
               "Default",
               "DevBuild",
               "ReleaseMaa",
-              "ReleaseMaaCore",
               "ReleaseMaaResource",
               "SetMaaChangeLog",
               "SetPackageBundled",
               "SetVersion",
-              "UseCleanArtifact",
-              "UseCleanRelease",
-              "UseCleanReleaseOta",
+              "UseClean",
               "UseCommitVersion",
               "UseMaaChangeLog",
-              "UseMaaCore",
               "UseMaaDevBundle",
-              "UseMaaLegacyBundle",
-              "UseMaaLegacyBundleOta",
+              "UseMaaRelease",
               "UseMaaResource",
               "UseMaaResourceChangeLog",
               "UsePublishArtifact",
@@ -100,20 +95,15 @@
               "Default",
               "DevBuild",
               "ReleaseMaa",
-              "ReleaseMaaCore",
               "ReleaseMaaResource",
               "SetMaaChangeLog",
               "SetPackageBundled",
               "SetVersion",
-              "UseCleanArtifact",
-              "UseCleanRelease",
-              "UseCleanReleaseOta",
+              "UseClean",
               "UseCommitVersion",
               "UseMaaChangeLog",
-              "UseMaaCore",
               "UseMaaDevBundle",
-              "UseMaaLegacyBundle",
-              "UseMaaLegacyBundleOta",
+              "UseMaaRelease",
               "UseMaaResource",
               "UseMaaResourceChangeLog",
               "UsePublishArtifact",
