+++ tools/MaaBuilder/Build.Job.Publish.cs
@@ -1,0 +1,50 @@
+﻿using Nuke.Common;
+using Nuke.Common.Tools.GitHub;
+using Octokit;
+
+namespace MaaBuilder;
+
+public partial class Build
+{
+    Target UsePublishArtifact => _ => _
+        .After(SetPackageBundled, SetMaaChangeLog)
+        .Produces(RootDirectory / "artifacts" / "*.zip")
+        .OnlyWhenStatic(() => GitHubActions != null);
+
+    Target UsePublishRelease => _ => _
+        .After(UsePublishArtifact)
+        .OnlyWhenStatic(() => GitHubActions != null)
+        .Executes(() =>
+        {
+            Information("在 GitHub Actions 中运行，执行发布 Release 任务");
+            Information($"当前 Actions：{Parameters.GhActionName}");
+
+            if (Parameters.GhActionName == ActionConfiguration.DevBuild)
+            {
+                Information($"DevBuild 跳过发布 Release");
+                return;
+            }
+
+            GitHubTasks.GitHubClient ??= new GitHubClient(new ProductHeaderValue(nameof(NukeBuild)))
+            {
+                Credentials = new Credentials(Parameters.GitHubPersonalAccessToken)
+            };
+
+            if (Parameters.GhActionName == ActionConfiguration.ReleaseMaa)
+            {
+                Information($"运行 ReleaseMaa 将在 {Parameters.MainRepo} 创建 Release {Parameters.GhTag}");
+
+                CreateGitHubRelease(Parameters.MainRepo, Parameters.CommitHashFull, Version);
+
+                return;
+            }
+
+            // ReSharper disable once InvertIf
+            if (Parameters.GhActionName == ActionConfiguration.ReleaseMaaResource)
+            {
+                Information($"运行 ReleaseMaaResource 将在 {Parameters.MaaResourceReleaseRepo} 创建 Release {Version}");
+
+                CreateGitHubRelease(Parameters.MaaResourceReleaseRepo, "main", Version);
+            }
+        });
+}
