+++ tools/MaaBuilder/MaaBuilder.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
-    <RootNamespace></RootNamespace>
+    <RootNamespace>MaaBuilder</RootNamespace>
     <NoWarn>CS0649;CS0169</NoWarn>
     <NukeRootDirectory>..\..</NukeRootDirectory>
     <NukeScriptDirectory>..\..</NukeScriptDirectory>
@@ -32,6 +32,12 @@
 
   <ItemGroup>
     <None Remove="..\..\.github\workflows\release-maa-core.yml" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Content Include="..\..\package-definition.json">
+      <Link>definition\package-definition.json</Link>
+    </Content>
   </ItemGroup>
 
 </Project>
