+++ tools/MaaBuilder/Models/IPackageConfiguration.cs
@@ -1,0 +1,3 @@
+﻿namespace MaaBuilder.Models;
+
+public interface IPackageConfiguration { }
