+++ tools/MaaBuilder/Models/Package.MaaBundleOta.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using System.Text.Json.Serialization;
+
+namespace MaaBuilder.Models;
+
+public record PackageMaaBundleOta : IPackageConfiguration
+{
+    [JsonPropertyName("exclude")]
+    public List<string> Exclude { get; set; }
+}
