+++ tools/MaaBuilder/Models/Package.MaaCore.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using System.Text.Json.Serialization;
+
+namespace MaaBuilder.Models;
+
+public record PackageMaaCore : IPackageConfiguration
+{
+    [JsonPropertyName("include")]
+    public List<string> Include { get; set; }
+}
