+++ tools/MaaBuilder/Models/Package.MaaDependency.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using System.Text.Json.Serialization;
+
+namespace MaaBuilder.Models;
+
+public record PackageMaaDependency : IPackageConfiguration
+{
+    [JsonPropertyName("include")]
+    public List<string> Include { get; set; }
+}
