+++ tools/MaaBuilder/Models/PackageTypes.cs
@@ -1,0 +1,10 @@
+namespace MaaBuilder.Models;
+
+public enum PackageTypes
+{
+    MaaBundle,
+    MaaBundleOta,
+    MaaCore,
+    MaaDependency,
+    MaaDependencyNoAvx
+}
